[move-prover] include the experimental code in df for verification
Verification of diem-framework experimental code is turned-off
previously due to symbolic links not being properly handled. This is
wrong and in fact, has caused some specifications to be violated.

This commit re-worked the move prover testsuite to include the
experimental code for verification.

The offending module is the newly refactored `DiemAccount.move` in
the experimental code. I'll leave it to the refactoring team to
double check.

Closes: #9805