[move-prover] Refine error diagnosis
This refines the output we produce on errors if the option `-t` (`--trace`) is set. Specifically, if an assertion fails, we print the values of sub-expressions of the condition in a new section "Related Bindings" in the error output. For examples, see the new test expectations for `trace.move`.

Some changes in baselines come from general improvements to the rendering procedure (e.g. skipping duplicate line traces).

Closes: #9810