[faucet] add cookie support to the rest client
This enables our rest server to provide stickiness, which in turn
ensures that if a user sends a bunch of transactions, they all land on
the same fullnode. The alternative is very bad, assuming 10 transactions
and 10 fullnodes, they could end up in 10 different fullnodes. So txn1
to fn1, txn2 to fn2, ... Only txn1 would make it upstream and eventually
to the validators, txn2 would only make it upstream after txn1 is
committed. This can easily result in long delays or timeouts.

This was validated by running against a test network with cookie-based
stickiness enabled. Without cookies we got 135 200s and 21 500s, with
cookies we got 790 200s and 0 500s.

Closes: #191