[vm] Introducing a delta write (#2057)
Introduced a new `WriteOp::Delta`. Delta is parametrised by a partial update
function, such as f(x) = + x or f(x) = - x, operating on 128-bit unsigned int,
and a limit (a postcondition that ensures the result of delta application does
not overflow).

Note: in the future, we would like to keep deltas private to executor, for
example by modifying `ChangeSet` or `TransactionOutput` instead.